Refactor parsing
Remove tacit inference before unary dot operators. Add ability to use various procedure keywords to get symbols in reflection. If nullary expressions with certain keywords have tacit commas, add it in situation stage only if it makes sense so that reflection is possible.

+++ include/rq/context.hpp
@@ -31,7 +31,7 @@ struct Token;
 struct Expression;
 struct Module;
 enum class Keyword : std::uint32_t;
-enum class Situation : std::uint_fast32_t;
+enum class Situation : std::uint_fast8_t;
 
 enum class LogType : std::underlying_type_t<llvm::SourceMgr::DiagKind> {
   ERROR = llvm::SourceMgr::DiagKind::DK_Error,
